backend/requirements.txt
Django==2.2.6
psycopg2-binary==2.8.3
djangorestframework==3.9.2
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0